Refactoring, performance improvements (#240)
This is a big PR that changes the `XPathMapper` in several significant ways.

## Replace proxy-approach with setter/field-based approach
Previously, mapping typees were supposed to be interfaces, and `XPathMapper` would then create a proxy-based implementation of the API that would hit the underlying `Document` every time a method annotated with `@XPathBinding` or `@XPathRoot` was called.

While this was a pretty simple API, it has significant performance drawbacks when using the proxy implementation, since every Method call would incur a significant amount of DOM walking.

To fix this, the mapper now expects mapping types to be classes, where setters or fields are annotated:

```java
@XPathRoot(defaultNamespace = "http://www.tei-c.org/ns/1.0")
public class MyType {

  private String documentType;

  @XPathBinding("/tei:TEI/@ana")
  private void setDocumentType(String documentType) {
    this.documentType = documentType.replaceAll("^#", "");
  }

  public String getDocumentType() {
    return documentType;
  }

  @XPathRoot
  MySubType subType;

  MySubType getSubType() {
    return subType
  }
}
```

To create an instance of the class from an XML Document, construct a mapper for the type and call its `readDocument` methtod:
```java
XPathMapper<MyType> mapper = new XPathMapper(MyType.class);
MyType obj = mapper.readDocument(doc);
```

## Make `XPathMapper<T>` re-usable and thread-safe
As could be seen in the above example, the constructor for `XPathMapper` now only takes as arguments the type and no longer the document. This is because with this PR, the implementation will perform all reflection-based dynamic code during construction of the mapper, and no longer at runtime. This means that you can now *reuse* the mapper for as many documents as you want, as long as you map them to the same type. The mapper is now also *thread-safe*, so you can just declare it as a Singleton in your application and use it from as many threads as you like.

## Refactoring of internals
The internals of the mapper also underwent significant changes. The code is now split between to main classes, `XPathMapper<T>` takes care of the Java-side of the mapping, i.e. reading out the annotations, making sanity checks on the types and setting the actual values on the target objects. All XML-related code has been moved to `DocumentReader`, which is constructed for each new document that needs to be mapped and takes care of using XPath to read out values from the XML.

## Commits

* Extend XPathMapper to work with classes instead of interfaces

* Checkstyle fixes

* Code cleanup

* Intermediate checkin, broke everything

* Rip out proxy logic

* Test fixes

* Test fixes

* Fixed another test

* Rewrote remaining tests

* xpathmapperr: Heavy refactoring

* Re-enabled previously disabled tests

* xpathmapper: Make XPathMapper<T> reusable

* checkstyle fixes

* javadoc: Disable doclint

* xpathmapper: Move reflection methods to mapper

Co-authored-by: Johannes Baiter <johannes.baiter@bsb-muenchen.de>